Update websocket implementation
tried to write a test for it but really was not working out
fastapi is NOT good for websockets at all right now
but maybe in a future version? it's still quite new
hopefully this implementation does not have any major bugs, but i guess ill find out when it comes to implementing it in the app.